+++ lib/brando/blueprint.ex
@@ -201,7 +201,6 @@ defmodule Brando.Blueprint do
       import Brando.Blueprint.Forms
       import Brando.Blueprint.Identifier.DSL
       import Brando.Blueprint.Listings
-      import Brando.Blueprint.Listings.Components
       import Brando.Blueprint.Naming
       import Brando.Blueprint.Relations
       import Brando.Blueprint.Trait