+++ lib/brando/content/module.ex
@@ -33,6 +33,7 @@ defmodule Brando.Content.Module do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   alias Phoenix.LiveView.JS