+++ lib/brando/content/palette.ex
@@ -16,6 +16,7 @@ defmodule Brando.Content.Palette do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   identifier "[{{ entry.namespace }}] {{ entry.name }}"