+++ lib/brando/content/table_template.ex
@@ -14,6 +14,7 @@ defmodule Brando.Content.TableTemplate do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   @type t :: %__MODULE__{}