+++ lib/brando/galleries/gallery.ex
@@ -11,6 +11,7 @@ defmodule Brando.Galleries.Gallery do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
   import Ecto.Query
 
   trait Brando.Trait.Timestamped