+++ lib/brando/images/image.ex
@@ -11,6 +11,7 @@ defmodule Brando.Images.Image do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   alias Brando.Images.Focal