+++ lib/brando/navigation/menu.ex
@@ -10,6 +10,7 @@ defmodule Brando.Navigation.Menu do
 
   use Gettext, backend: Brando.Gettext
 
+  import Brando.Blueprint.Listings.Components
   import Ecto.Query
 
   alias Brando.Content.Var