+++ lib/brando/pages/page.ex
@@ -12,6 +12,7 @@ defmodule Brando.Pages.Page do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   alias Brando.JSONLD
   alias Brando.Pages