+++ lib/brando/sites/global_set.ex
@@ -9,6 +9,7 @@ defmodule Brando.Sites.GlobalSet do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   trait Brando.Trait.Creator
   trait Brando.Trait.CastPolymorphicEmbeds