+++ lib/brando/users/user.ex
@@ -13,6 +13,7 @@ defmodule Brando.Users.User do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   alias Brando.Users.UserConfig