+++ lib/brando/videos/video.ex
@@ -11,6 +11,7 @@ defmodule Brando.Videos.Video do
     gettext_module: Brando.Gettext
 
   use Gettext, backend: Brando.Gettext
+  import Brando.Blueprint.Listings.Components
 
   trait Brando.Trait.Creator
   trait Brando.Trait.Timestamped